Panic when AWS querying errors
Have the stats functions panic if queried immediately after updating the
stats have failed. Have Beta() panic on error from ECS.

This is so these panics can be handled from the querying control. The
Start() function of the control, asynchronous, is now replaced by Run()
which must be handled synchronously by the caller and can panic().