Simplify beta integral
Remove the rather complex integration system that was in place for the
beta integral, given that's always done in fixed intervals (the same as
the way Y was got from its derivative).

Since Y and the beta integral are only present as a ratio, remove the
multiplication by the period.