Support for internal concurrency in workers
The logic for estimating R up when the system is starved assumes no
internal concurrency for the workers, so it estimates the number of
active workers with the number of messages currently being worked on.

This fails in an multithreading (internally concurrent) workers
scenario, leading to the control process never adjusting R up (or taking
much longer than needed).

Fix this by keeping a measure of the internal concurrency by sampling
the ratio of (X - Y - Q) / β, which will be larger than the unit when
workers are internally concurrent.