Consider Q <= β as Q ≈ 0
In rapidly changing enviroments, i.e. high DX and DY, it can be hard to
spot an actual Q = 0 so a system that's consistently overscaled can keep
that state without scaling in, getting Q = 1, Q = 2 etc..

Consider Q > β only as a queued-up system, which makes scaling in
possible in these environments, without affecting too negatively
low-paced environments, for example β ≈ 1 -- in these considering Q = 1
as overscaled might be a mistake but it will be corrected when the
system queues up (Q > 1) in the following control period.